ReguGuard-AI/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openai>=2.21.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
]
